+++ examples/create_shoe_index.php
@@ -22,7 +22,9 @@ foreach ($data as $item) {
     $itemData = array_merge($itemData, $item['features']);
     $searchIndex->addRecord($recordId, $itemData);
 }
+$searchIndex->optimize();
 // save index data to some storage
 $indexData = $searchIndex->getData();
 // We will use file for example
 file_put_contents('./data/shoe-index.json', json_encode($indexData));
+echo 'Index created' . PHP_EOL;